Update show.html.erb
Changing the 'View Items' path to merchant_items_path, instead of the path a merchant would be routed (merchant_user_items_path) resolves the 404 page issue. 

The Admin can now link to a view of a Merchant's active items from the Merchant dashboard view, but cannot add, edit, delete, or activate/deactivate an item as a Merchant has the authorization to do. I assume this is correct.